fix some russian locales; now locales are emded in binary
