+++ .github/workflows/deploy.yml
@@ -8,7 +8,7 @@ on:
       - "cmd/**"
       - "config/**"
       - "handlers/**"
-      - "bundle_i18n/**"
+      - "locales/**"
       - "repository/**"
       - "format/**"
       - "middleware/**"
