+++ locales/messages.ru.json
@@ -91,12 +91,7 @@
       "messageOverflow": "<b>сообщение отредактировано</b>\nчат: {{.ResolvedChatName}}\n<i>ID чата: {{.ChatID}}</i>\n<i>время редактирования: {{.Date}}</i>\nСлишком много деталей, пожалуйста, смотрите детали по кнопке ниже...",
       "buttons": {
         "log": "получить JSON",
-        "getFile": {
-          "one": "получить {{.Count}} файл",
-          "few": "получить {{.Count}} файла",
-          "many": "получить {{.Count}} файлов",
-          "other": "получить {{.Count}} файлов"
-        }
+        "getFile": "получить файл"
       },
       "text": {
         "added": "📝 <b>текст добавлен:</b> {{.New}}",
